Circular buffers

+++ demo1.py
@@ -5,9 +5,7 @@ from scipy.io import wavfile
 import sys
 sys.path.insert(1, 'scripts and data') 
 import mp3
-import frame
-import nothing
-from subband_filtering import coder0, decoder0, codec0
+from subband_filtering import codec0
 
 h = np.load("scripts and data//h.npy", allow_pickle=True).tolist()['h'].reshape(-1, )
 M = 32
@@ -41,7 +39,7 @@ samplerate, wavin = wavfile.read("scripts and data//myfile.wav")
 xhat, Ytot = codec0(wavin, h, M, N)
 
 
-wavfile.write("xhat.wav", samplerate, xhat)
+wavfile.write("xhat.wav", samplerate, np.int16(xhat))
 
 
 plt.plot(xhat)
